Border system
- Tiles not in borders can't be claimed
- Settlements claim more tiles.
- AI uses borders.

+++ Source/1.3/Resources/ResourceModifier.cs
@@ -24,10 +24,10 @@ namespace Empire_Rewritten
     public struct ResourceModifier
     {
         public ResourceDef def;
-        public int offset;
+        public float offset;
         public float multiplier;
 
-        public ResourceModifier(ResourceDef resourceDef, int offsetValue = 0, float multiplierValue = 1)
+        public ResourceModifier(ResourceDef resourceDef, float offsetValue = 0, float multiplierValue = 1)
         {
             def = resourceDef;
             offset = offsetValue;
